# api/docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.CustomPresetForm:
    properties:
      app_name:
        type: string
      custom_json:
        type: string
      id:
        type: integer
      name:
        type: string
      platform:
        type: string
      version:
        type: string
    required:
    - name
    - platform
    - version
    type: object
...
  model.CustomPresetSafe:
    properties:
      app_name:
        type: string
      created_at:
        type: string
      custom_json:
        type: string
      id:
        type: integer
      name:
        type: string
      platform:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      version:
        type: string
    type: object
  model.CustomPresetSafeList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.CustomPresetSafe'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: API
  version: "1.0"
paths:
...
  /admin/config/admin:
    get:
      description: Public pre-authentication route returning the admin title and,
        when a valid api-token is supplied, the personalized greeting.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get public admin configuration
      tags:
      - ADMIN
  /admin/config/all:
    get:
      description: Admin-only configuration used by the admin panel.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Get all admin configuration
      tags:
      - ADMIN
  /admin/config/app:
    get:
      consumes:
      - application/json
      description: APP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: APP
      tags:
      - ADMIN
  /admin/config/server:
    get:
      consumes:
      - application/json
      description: ',webclientapi-server'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: RUSTDESK
      tags:
      - ADMIN
...
  /admin/custom_preset/create:
    post:
      consumes:
      - application/json
      description: Administrators with AdminPrivilege may save a custom preset; provider-derived
        build references are resolved server-side.
      parameters:
      - description: Custom-preset payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CustomPresetForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.CustomPresetSafe'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Create a custom preset
      tags:
      - CustomPreset
  /admin/custom_preset/delete:
    post:
      consumes:
      - application/json
      description: Administrators with AdminPrivilege may delete their own custom
        preset.
      parameters:
      - description: Custom-preset identifier
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CustomPresetForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Delete a custom preset
      tags:
      - CustomPreset
  /admin/custom_preset/detail/{id}:
    get:
      description: Administrators with AdminPrivilege may read their own redacted
        custom preset.
      parameters:
      - description: Custom preset ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.CustomPresetSafe'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Get a custom preset
      tags:
      - CustomPreset
  /admin/custom_preset/list:
    get:
      description: Administrators with AdminPrivilege receive a paginated redacted
        custom-preset list.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of presets per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.CustomPresetSafeList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: List custom presets
      tags:
      - CustomPreset
  /admin/custom_preset/update:
    post:
      consumes:
      - application/json
      description: Administrators with AdminPrivilege may update their own custom
        preset; provider-derived build references are resolved server-side.
      parameters:
      - description: Custom-preset update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CustomPresetForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.CustomPresetSafe'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Update a custom preset
      tags:
      - CustomPreset
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"